TFGALL/requirements.txt
# Uso (dentro del entorno virtual): pip install -r requirements.txt
absl-py==1.4.0
altgraph==0.17.3
appnope==0.1.3
asttokens==2.2.1
Babel==2.12.1
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
customtkinter==5.1.2
cycler==0.11.0
darkdetect==0.8.0
decorator==5.1.1
executing==1.2.0
filelock==3.11.0
fonttools==4.39.3
google-auth==2.17.2
google-auth-oauthlib==1.0.0
grpcio==1.53.0
idna==3.4
importlib-metadata==6.3.0
importlib-resources==5.12.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
macholib==1.16.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
modulegraph==0.19.3
mpmath==1.3.0
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pandas==2.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==4.21.2
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py2app==0.28.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
tkcalendar==1.6.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
vega-datasets==0.9.0
wcwidth==0.2.6
Werkzeug==2.2.3
XlsxWriter==3.0.9
zipp==3.15.0
